Remove non-code generation example
