+++ nmdc_dp_utils/llm/biosample_mapping/instructions.py
@@ -1,42 +1,61 @@
 system_prompt = '''
 # TASK
-Map raw mass spectrometry files to biosamples and processed samples by analyzing file naming conventions and material processing protocols.
+Generate a Python script that maps raw mass spectrometry files to biosamples and processed samples.
 
-# WORKFLOW
-1. Analyze input data (biosample attributes, raw files, material processing YAML)
-2. Generate CSV mapping based on file naming patterns
-3. Call `validate_biosample_mapping` tool with your CSV (if available)
-4. If validation fails, fix errors and revalidate
-5. Provide final CSV (CSV only, no markdown/commentary)
+# APPROACH
+Write a Python script that:
+1. Parses file naming patterns to extract relevant metadata
+2. Matches patterns to biosample information
+3. Determines the appropriate protocol and processed sample
+4. Generates a complete CSV mapping
 
 # INPUT DATA
 You will receive:
-1. **Biosample Attributes CSV**: Contains biosample metadata (id, name, samp_name)
-2. **Raw Files CSV**: Lists all raw mass spectrometry files (file_name)
-3. **Material Processing YAML**: Describes the laboratory processing steps and processed sample outputs
+- Biosample attributes (id, name)
+- Raw file names (list of files to map)
+- Material processing YAML (protocol steps and processed samples)
+- Additional context (optional): Information about naming conventions, experimental methods, and sample processing
+- Example mappings showing successful patterns
 
-# OUTPUT FORMAT
-CSV with columns:
-- `raw_data_identifier`: Raw file name
-- `biosample_id`: NMDC ID (nmdc:bsm-XX-XXXXXXXX), directly from biosample attributes CSV
-- `biosample_name`: Biosample name, must match the name in biosample attributes CSV for the given biosample_id
-- `match_confidence`: `high`, `medium`, or `low`
-- `processedsample_placeholder`: Processed sample ID from YAML that logically corresponds to the raw file based on the protocol steps
-- `material_processing_protocol_id`: Protocol name from YAML that describes the processing steps for this sample (e.g. "polar", "nonpolar")
+# OUTPUT
+Generate a complete, executable Python script that creates the output CSV.
 
-# MAPPING LOGIC
+# SCRIPT STRUCTURE
+```python
+import pandas as pd
+import re
 
-## 1. Parse Raw File Names
-File names encode: sample ID (S16, S1, etc.), replicate (A/B/C), timepoint (D45), method (HILIC=polar, C18/RP=nonpolar), ionization (POS/NEG)
+# Load input data
+biosamples = pd.read_csv('INPUT_BIOSAMPLE_PATH')
+files = pd.read_csv('INPUT_FILES_PATH')
 
-## 2. Match to Biosample
-Match sample code, timepoint, replicate to biosample `name`. Confidence: `high` (all match), `medium` (missing one), `low` (partial match)
+def map_file_to_biosample(filename):
+    """Extract metadata and match to biosample"""
+    # Analyze filename patterns
+    # Match to biosamples
+    # Determine protocol from evidence in filename
+    pass
 
-## 3. Determine Processed Sample and Protocol
-Match file method to protocol: HILIC/hydrophilic→polar protocol, C18/RP→nonpolar. Use final ProcessedSample in protocol (highest numbered, represents MS-ready sample)
+# Process all files
+results = []
+for filename in files['COLUMN_NAME']:
+    mapping = map_file_to_biosample(filename)
+    results.append({...})
 
-## 4. Edge Cases
-- Same biosample may have multiple files (replicates, POS/NEG modes)
-- QC/blank/ExCtrl files: may have empty biosample_id
-- Low confidence matches: provide best guess, mark as `low`
+# Save output
+output = pd.DataFrame(results)
+output.to_csv('OUTPUT_PATH', index=False)
+```
+
+# KEY POINTS
+- Study the example mappings to understand the patterns
+- Look for clues in filenames (sample IDs, methods, timepoints, replicates)
+- Match sample identifiers to biosample names
+- Use analytical method indicators (HILIC, C18, RP, etc.) to determine protocols
+- Handle QC/blank/control samples appropriately (may have empty biosample fields)
+- Set confidence based on match quality
+
+# OUTPUT REQUIREMENTS
+Provide ONLY the Python script code. No markdown blocks, no explanations outside script comments.
+The script must be complete and ready to execute.
 '''
